requirements.txt
pytest == 7.2.1
faker == 16.6.1
selenium == 4.8.0
allure-pytest == 2.12.0